refactor: add form and token contexts in create vesting to get data back when quitting the signing page
