+++ apps/create-vesting/src/app/layout.tsx
@@ -2,6 +2,8 @@
 
 import "@bleu/cow-hooks-ui/global.css";
 import { IFrameContextProvider } from "@bleu/cow-hooks-ui";
+import { TokenContextProvider } from "#/context/token";
+import { FormContextProvider } from "#/context/form";
 import Head from "next/head";
 import type * as React from "react";
 import "@fortawesome/fontawesome-svg-core/styles.css";
@@ -15,11 +17,15 @@ export default function Layout({ children }: { children: React.ReactNode }) {
         <link rel="manifest" href="/manifest.json" />
       </Head>
       <IFrameContextProvider>
-        <body className="bg-transparent">
-          <div className="font-sans font-normal scrollbar-w-1 scrollbar scrollbar-thumb-color-paper-darkest scrollbar-track-color-paper-darker h-screen overflow-y-scroll">
-            {children}
-          </div>
-        </body>
+        <TokenContextProvider>
+          <body className="bg-transparent">
+            <FormContextProvider>
+              <div className="font-sans font-normal scrollbar-w-1 scrollbar scrollbar-thumb-color-paper-darkest scrollbar-track-color-paper-darker h-screen overflow-y-scroll">
+                {children}
+              </div>
+            </FormContextProvider>
+          </body>
+        </TokenContextProvider>
       </IFrameContextProvider>
     </html>
   );
